// web-ui/client/gen/sales/taskTypeEnum.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `call` - Call
* `email` - Email
* `meeting` - Meeting
* `demo` - Demo
 */
export type TaskTypeEnum = typeof TaskTypeEnum[keyof typeof TaskTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaskTypeEnum = {
  call: 'call',
  email: 'email',
  meeting: 'meeting',
  demo: 'demo',
} as const;