// web-ui/client/gen/sales/taskStatusEnum.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `pending` - Pending
* `completed` - Completed
* `overdue` - Overdue
 */
export type TaskStatusEnum = typeof TaskStatusEnum[keyof typeof TaskStatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaskStatusEnum = {
  pending: 'pending',
  completed: 'completed',
  overdue: 'overdue',
} as const;